Convert camera snapshot to image entity
